Update README.md
Added pinout, connections and pictures. <p align="center"> which center the content inside the paragraph. <img> is an HTML tag used to display an image. src="..." tells it where to get the image from — it's the image URL. height="200" sets the height of the image to 200 pixels. No width is specified, so the browser automatically calculates the width to keep the image's natural shape (aspect ratio). The /> closes the <img> tag (self-closing tag).